+++ src/features/menu/index.ts
@@ -1,0 +1,2 @@
+export { default as AppMenu } from './AppMenu.svelte';
+export type { MenuItem } from './menu-item';