+++ src/widgets/chart-canvas/index.ts
@@ -1,1 +1,2 @@
 export { default as ChartCanvas } from './ChartCanvas.svelte';
+export type { MapCommands } from './commands';